+++ include/databento/detail/buffer.hpp
@@ -80,7 +80,7 @@ class Buffer : public IReadable, public IWritable {
 
   UniqueBufPtr buf_;
   std::byte* end_;
-  std::byte* read_pos_{};
-  std::byte* write_pos_{};
+  std::byte* read_pos_;
+  std::byte* write_pos_;
 };
 }  // namespace databento::detail