+++ tests/src/http_client_tests.cpp
@@ -1,12 +1,11 @@
+#include <gmock/gmock.h>
 #include <gtest/gtest.h>
 #include <nlohmann/json.hpp>
-
-#include <algorithm>
-#include <iostream>
 
 #include "databento/detail/http_client.hpp"
 #include "databento/log.hpp"
 #include "mock/mock_http_server.hpp"
+#include "mock/mock_log_receiver.hpp"
 
 namespace databento::detail::tests {
 class HttpClientTests : public ::testing::Test {
@@ -22,16 +21,21 @@ TEST_F(HttpClientTests, TestLogWarnings) {
       "Warning: Large request"};
   mock_server_.MockGetJson("/warn", {}, {}, warnings);
   const auto port = mock_server_.ListenOnThread();
-  HttpClient target{ILogReceiver::Default(), kApiKey, "localhost",
+  databento::tests::mock::MockLogReceiver mock_logger{
+      [](auto call_count, databento::LogLevel level, const std::string& msg) {
+        EXPECT_EQ(level, LogLevel::Warning);
+        if (call_count == 0) {
+          EXPECT_THAT(
+              msg,
+              testing::EndsWith(
+                  "Server DeprecationWarning: stype product_id is deprecated"));
+        } else {
+          EXPECT_THAT(msg, testing::EndsWith("Server Warning: Large request"));
+        }
+      }};
+  HttpClient target{&mock_logger, kApiKey, "localhost",
                     static_cast<std::uint16_t>(port)};
-  testing::internal::CaptureStderr();
   target.GetJson("/warn", {});
-  std::clog.flush();
-  const std::string output = testing::internal::GetCapturedStderr();
-  EXPECT_EQ(std::count(output.begin(), output.end(), '\n'), 2);
-  EXPECT_NE(
-      output.find("Server DeprecationWarning: stype product_id is deprecated"),
-      std::string::npos);
-  EXPECT_NE(output.find("Server Warning: Large request"), std::string::npos);
+  ASSERT_EQ(mock_logger.CallCount(), 2);
 }
 }  // namespace databento::detail::tests